jumga-backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
beautifulsoup4==4.9.3
cachetools==4.2.0; python_version ~= '3.5'
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==3.3.1
defusedxml==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==0.5.0
django-allauth==0.44.0
django-cors-headers==3.6.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages[google]==1.11.1
django-taggit==1.3.0
django==3.1.5
djangorestframework==3.12.2; python_version >= '3.5'
google-api-core==1.25.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth==1.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-core==1.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-storage==1.35.0
google-crc32c==1.1.1; python_version >= '3.5'
google-resumable-media==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gunicorn==20.0.4
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
lxml==4.6.2
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.0.1; python_version >= '3.6'
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7; python_version >= '3.6'
six==1.15.0
soupsieve==2.1; python_version >= '3.0'
sqlparse==0.4.1; python_version >= '3.5'
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==5.2.0
